Fix type mismatch in pg_show_plans.pid and level
This was causing segfaults on 32-bit Linux.

+++ pg_show_plans--1.1.sql
@@ -35,8 +35,8 @@ AS 'MODULE_PATHNAME'
 LANGUAGE C;
 
 CREATE FUNCTION pg_show_plans(
-	OUT pid int8,
-	OUT level int8,
+	OUT pid int,
+	OUT level int,
 	OUT userid oid,
 	OUT dbid oid,
 	OUT plan text
